refactor folder structure and rename files
